Vector Sets: harden VSIM_execute against negative hnsw_search return
hnsw_search() / hnsw_search_with_filter() / hnsw_ground_truth_with_filter()
all return int and may return -1 on error (e.g. k == 0, allocation
failure). The result was stored in an unsigned int, so -1 coerced to
UINT_MAX and drove the result loop past the end of the neighbors/distances
allocations.

Make 'found' signed, clamp negative returns to 0, and update the loop
iterator to match. The negative return path is not currently reachable
from a healthy client, but the type signature should hold up against
future callers and any non-VREM way of producing a degenerate search
state.

See MOD-16035.

(cherry picked from commit 124d206b19021da572b22ad6837ab767987ffa20)

+++ modules/vector-sets/vset.c
@@ -838,7 +838,7 @@ void VSIM_execute(RedisModuleCtx *ctx, struct vsetObject *vset,
     /* Perform search */
     hnswNode **neighbors = RedisModule_Alloc(sizeof(hnswNode*)*ef);
     float *distances = RedisModule_Alloc(sizeof(float)*ef);
-    unsigned int found;
+    int found;
     if (ground_truth) {
         found = hnsw_ground_truth_with_filter(vset->hnsw, vec, ef, neighbors,
                     distances, slot, 0,
@@ -865,7 +865,7 @@ void VSIM_execute(RedisModuleCtx *ctx, struct vsetObject *vset,
         RedisModule_ReplyWithArray(ctx, REDISMODULE_POSTPONED_LEN);
 
     long long arraylen = 0;
-    for (unsigned int i = 0; i < found && i < count; i++) {
+    for (int i = 0; i < found && (unsigned long)i < count; i++) {
         if (distances[i]/2 > epsilon) break;
         struct vsetNodeVal *nv = neighbors[i]->value;
         RedisModule_ReplyWithString(ctx, nv->item);
